+++ project/app-infrastructure/src/main/java/mobi/lab/sample/infrastructure/auth/AuthProvider.kt
@@ -8,7 +8,7 @@ import mobi.lab.sample.infrastructure.auth.remote.ApiSessionMapper
 import mobi.lab.sample.infrastructure.auth.remote.AuthResource
 import javax.inject.Inject
 
-internal class AuthProvider @Inject constructor(
+class AuthProvider @Inject constructor(
     private val res: AuthResource,
     private val sessionMapper: ApiSessionMapper
 ) : AuthGateway {