+++ project/app-infrastructure/src/main/java/mobi/lab/sample/infrastructure/auth/local/DbSessionMapper.kt
@@ -3,7 +3,7 @@ package mobi.lab.sample.infrastructure.auth.local
 import mobi.lab.sample.domain.entities.Session
 import javax.inject.Inject
 
-internal class DbSessionMapper @Inject constructor() {
+class DbSessionMapper @Inject constructor() {
 
     fun toEntity(item: DbSession): Session {
         return Session(item.token)