+++ project/app-infrastructure/src/main/java/mobi/lab/sample/infrastructure/auth/remote/ApiSessionMapper.kt
@@ -3,7 +3,7 @@ package mobi.lab.sample.infrastructure.auth.remote
 import mobi.lab.sample.domain.entities.Session
 import javax.inject.Inject
 
-internal class ApiSessionMapper @Inject constructor() {
+class ApiSessionMapper @Inject constructor() {
 
     fun toEntity(item: ApiSession): Session {
         return Session(item.token)